finished models and continued working on routes

+++ routes/api/product-routes.js
@@ -6,34 +6,38 @@ const { Product, Category, Tag, ProductTag } = require('../../models');
 // get all products
   // find all products
   // be sure to include its associated Category and Tag data***
-  router.get('/', async (req, res) => {
-    try {
-      const productData = await Product.findAll();
-      res.status(200).json(productData);
-    } catch (err) {
-      res.status(500).json(err);
-    }
+  router.get('/', (req, res) => {
+    Product.findAll({
+      include:[
+        Category,
+        {
+          model: Tag,
+          through: ProductTag
+        }
+      ]
+    })
+    .then((products)=>res.json(products))
+    .catch((err)=>res.status(500).json(err))
   });
   
-
 // get one product
   // find a single product by its `id`
   // be sure to include its associated Category and Tag data
-  router.get('/:id', async (req, res) => {
-    try {
-      const productData = await Product.findByPk(req.params.id, {
-        include: [{ model: Category, through: Tag, as: 'product_category' }]
-      });
-  
-      if (!locationData) {
-        res.status(404).json({ message: 'No location found with this id!' });
-        return;
-      }
-  
-      res.status(200).json(locationData);
-    } catch (err) {
-      res.status(500).json(err);
-    }
+  router.get('/:id', (req, res) => {
+    Product.findOne({
+      where:{
+        id:req.params.id
+      },
+      include:[
+        Category,
+        {
+          model: Tag,
+          through: ProductTag
+        }
+      ]
+    })
+    .then((product)=>res.json(product))
+    .catch((err)=>res.status(400).json(err))
   });
 
 // create new product
@@ -49,7 +53,7 @@ router.post('/', (req, res) => {
   Product.create(req.body)
     .then((product) => {
       // if there's product tags, we need to create pairings to bulk create in the ProductTag model
-      if (req.body.tagIds.length) {
+      if (req.body.tagIds && req.body.tagIds.length) {
         const productTagIdArr = req.body.tagIds.map((tag_id) => {
           return {
             product_id: product.id,
@@ -77,57 +81,45 @@ router.put('/:id', (req, res) => {
     },
   })
     .then((product) => {
+      if(req.body.tagIds && req.body.tagIds.length){
+        const productTags = ProductTag.findAll({ where: { product_id: req.params.id } });
+        const productTagIds = productTags.map(({ tag_id }) => tag_id);
+        // create filtered list of new tag_ids
+        const newProductTags = req.body.tagIds
+          .filter((tag_id) => !productTagIds.includes(tag_id))
+          .map((tag_id) => {
+            return {
+              product_id: req.params.id,
+              tag_id,
+            };
+          });
+        // figure out which ones to remove
+        const productTagsToRemove = productTags
+          .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
+          .map(({ id }) => id);
+  
+        // run both actions
+        return Promise.all([
+          ProductTag.destroy({ where: { id: productTagsToRemove } }),
+          ProductTag.bulkCreate(newProductTags),
+        ]);
+      }
+      return res.json(product)
+      })
+      .catch((err)=>res.status(400).json(err))
       // find all associated tags from ProductTag
-      return ProductTag.findAll({ where: { product_id: req.params.id } });
-    })
-    .then((productTags) => {
-      // get list of current tag_ids
-      const productTagIds = productTags.map(({ tag_id }) => tag_id);
-      // create filtered list of new tag_ids
-      const newProductTags = req.body.tagIds
-        .filter((tag_id) => !productTagIds.includes(tag_id))
-        .map((tag_id) => {
-          return {
-            product_id: req.params.id,
-            tag_id,
-          };
-        });
-      // figure out which ones to remove
-      const productTagsToRemove = productTags
-        .filter(({ tag_id }) => !req.body.tagIds.includes(tag_id))
-        .map(({ id }) => id);
-
-      // run both actions
-      return Promise.all([
-        ProductTag.destroy({ where: { id: productTagsToRemove } }),
-        ProductTag.bulkCreate(newProductTags),
-      ]);
-    })
-    .then((updatedProductTags) => res.json(updatedProductTags))
-    .catch((err) => {
-      // console.log(err);
-      res.status(400).json(err);
-    });
+      // return ProductTag.findAll
 });
 
 // delete one product by its `id` value
-router.delete('/:id', async (req, res) => {
-  try {
-    const productData = await Product.destroy({
-      where: {
-        id: req.params.id
-      }
-    });
-
-    if (!productData) {
-      res.status(404).json({ message: 'No product found with this id!' });
-      return;
+router.delete('/:id', (req, res) => {
+  Product.destroy({
+    where: {
+      id:req.params.id
     }
-
-    res.status(200).json(productData);
-  } catch (err) {
-    res.status(500).json(err);
-  }
+  })
+  .then((products)=>res.json(products))
+  .catch((err)=>res.status(400).json(err))
 });
 
 module.exports = router;
